backend/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
google-generativeai
httpx
pydantic
python-dotenv
livekit-api
parallel-web
